Move GL utilities out to their own library
