+++ libraries/entities-renderer/src/RenderableWebEntityItem.cpp
@@ -18,7 +18,7 @@
 #include <DeferredLightingEffect.h>
 #include <GeometryCache.h>
 #include <PerfStat.h>
-#include <OffscreenQmlSurface.h>
+#include <gl/OffscreenQmlSurface.h>
 #include <AbstractViewStateInterface.h>
 #include <GLMHelpers.h>
 #include <PathUtils.h>
