+++ tests/render-utils/CMakeLists.txt
@@ -6,6 +6,6 @@ setup_hifi_project(Quick Gui OpenGL)
 set_target_properties(${TARGET_NAME} PROPERTIES FOLDER "Tests/manual-tests/")
 
 # link in the shared libraries
-link_hifi_libraries(render-utils gpu shared)
+link_hifi_libraries(render-utils gl gpu shared)
 
 copy_dlls_beside_windows_executable()
